refactor(vscode): extract shared diff review UI
