+++ packages/kilo-vscode/tests/unit/agent-manager-arch.test.ts
@@ -23,14 +23,14 @@ const TSX_FILES = [
   path.join(ROOT, "webview-ui/agent-manager/NewWorktreeDialog.tsx"),
   path.join(ROOT, "webview-ui/agent-manager/sortable-tab.tsx"),
   path.join(ROOT, "webview-ui/agent-manager/DiffPanel.tsx"),
-  path.join(ROOT, "webview-ui/agent-manager/FullScreenDiffView.tsx"),
-  path.join(ROOT, "webview-ui/agent-manager/MarkdownDiffView.tsx"),
-  path.join(ROOT, "webview-ui/agent-manager/MarkdownAnnotationLayer.tsx"),
-  path.join(ROOT, "webview-ui/agent-manager/markdown-comment-ranges.ts"),
-  path.join(ROOT, "webview-ui/agent-manager/DiffEndMarker.tsx"),
-  path.join(ROOT, "webview-ui/agent-manager/FileTree.tsx"),
-  path.join(ROOT, "webview-ui/agent-manager/review-annotations.ts"),
-  path.join(ROOT, "webview-ui/agent-manager/review-annotation-speech.tsx"),
+  path.join(ROOT, "webview-ui/diff-viewer/FullScreenDiffView.tsx"),
+  path.join(ROOT, "webview-ui/diff-viewer/MarkdownDiffView.tsx"),
+  path.join(ROOT, "webview-ui/diff-viewer/MarkdownAnnotationLayer.tsx"),
+  path.join(ROOT, "webview-ui/diff-viewer/markdown-comment-ranges.ts"),
+  path.join(ROOT, "webview-ui/diff-viewer/DiffEndMarker.tsx"),
+  path.join(ROOT, "webview-ui/diff-viewer/FileTree.tsx"),
+  path.join(ROOT, "webview-ui/diff-viewer/review-annotations.ts"),
+  path.join(ROOT, "webview-ui/diff-viewer/review-annotation-speech.tsx"),
   path.join(ROOT, "webview-ui/agent-manager/MultiModelSelector.tsx"),
   path.join(ROOT, "webview-ui/agent-manager/ApplyDialog.tsx"),
   path.join(ROOT, "webview-ui/agent-manager/WorktreeItem.tsx"),
