+++ packages/kilo-vscode/webview-ui/diff-viewer/DiffEndMarker.tsx
