+++ packages/kilo-vscode/webview-ui/diff-viewer/DiffViewerApp.tsx
@@ -11,8 +11,8 @@ import { File } from "@kilocode/kilo-ui/file"
 import { Icon } from "@kilocode/kilo-ui/icon"
 import { ThemeProvider } from "@kilocode/kilo-ui/theme"
 import { Toast } from "@kilocode/kilo-ui/toast"
-import { FullScreenDiffView } from "../agent-manager/FullScreenDiffView"
-import { mergeWorktreeDiffs } from "../agent-manager/diff-state"
+import { FullScreenDiffView } from "./FullScreenDiffView"
+import { mergeWorktreeDiffs } from "./diff-state"
 import { LanguageProvider, useLanguage } from "../src/context/language"
 import { ServerProvider, useServer } from "../src/context/server"
 import { ConfigProvider } from "../src/context/config"
