+++ packages/kilo-vscode/webview-ui/diff-viewer/FileTree.tsx
