+++ packages/kilo-vscode/webview-ui/diff-viewer/MarkdownAnnotationLayer.tsx
