+++ packages/kilo-vscode/webview-ui/diff-viewer/diff-open-policy.ts
