+++ packages/kilo-vscode/webview-ui/diff-viewer/diff-state.ts
