+++ packages/kilo-vscode/webview-ui/diff-viewer/file-tree-utils.ts
