+++ packages/kilo-vscode/webview-ui/diff-viewer/markdown-comment-ranges.ts
